+++ charms/garm/src/garm_client/test/test_create_template_params.py
@@ -35,9 +35,7 @@ class TestCreateTemplateParams(unittest.TestCase):
         model = CreateTemplateParams()
         if include_optional:
             return CreateTemplateParams(
-                data = [
-                    56
-                    ],
+                data = '',
                 description = '',
                 forge_type = '',
                 name = '',